automation/requirements.txt
requests>=2.28.0
Pillow>=9.0.0
python-dotenv>=0.19.0
schedule>=1.1.0
pytz>=2021.3